requirements.txt
tensorflow
tensorflow_text
tensorflow_datasets
pydantic
pydantic-dotenv
nltk
Jinja